+++ packages/ui/react-ui-stack/src/components/PlankHeading.tsx
@@ -29,7 +29,7 @@ import { useAttention } from '@dxos/react-ui-attention';
 import { descriptionText, mx } from '@dxos/react-ui-theme';
 import { getHostPlatform } from '@dxos/util';
 
-import { StackItemResizeHandle } from './StackItem';
+import { StackItemResizeHandle } from './StackItemResizeHandle';
 import { translationKey } from '../translations';
 
 // TODO(thure): Dedupe (also in react-ui-navtree)
